index.md: C Conf in San Diego cancelled
Thank you for your support of C Conf. The community support was crucial
in motivating the work to get C Conf 2012 in San Diego as far as it did.

Regretfully, C Conf in San Diego on August 28th, 2012 is canceled.

I apologize for the problems this causes. Please send your questions or
concerns to contact@cconf.org

The core reason for cancelling was a failure to develop a full line-up
of presentations. There was a great set of proposed speakers and ideas
but a combination of timing conflicts, location and low proposal count
made it impossible to get enough speakers to provide you with a great
experience.

In the near future a series of smaller evening events will be organized
around the San Francisco/Bay Area on topics around the C language based
on ideas from the call for papers. Pay attention to the [announce
mailing][announce] list for more details.

[announce]: https://groups.google.com/forum/?fromgroups#!forum/cconf-announce

Thanks again.